+++ src/types/components/molecules/list.ts
@@ -1,7 +1,6 @@
 import {
   BuilderComponent,
   ClickableComponent,
-  ComponentWithCallback,
   MoBrixUiComponent,
 } from "../../../types/global";