+++ src/types/components/organisms/date-picker.ts
@@ -21,6 +21,7 @@ export type DatePickerProps = ComponentWithCallback<CalendarDate> &
   ComponentWithValue<CalendarDate> &
   ClosableComponent &
   CalendarProps & {
+    /** MoBrix-ui sharedProps applied on the internal calendar */
     calendarProps?: CommonProps;
   };